pyproject.toml
[project]
name = "derive"
version = "0.1.0"
description = "Derivé — step-by-step mathematical derivation tool with symbolic verification"
readme = "README.md"
requires-python = ">=3.11,<3.12"
authors = [
    { name = "Khan Easa", email = "easabingias@gmail.com" }
]
license = { text = "MIT" }
dependencies = [
    "anthropic>=0.96.0",
    "python-dotenv>=1.2.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend/src/derive"]

[dependency-groups]
dev = [
    "mypy>=1.20.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.11",
]

[tool.ruff]
# Python version ruff assumes when checking
target-version = "py311"

# Maximum line length
line-length = 100

# Directories to look at for linting and formatting
src = ["backend/src"]

# Folders to exclude entirely
extend-exclude = [
    ".venv",
    "frontend",
    "corpus",
]

[tool.ruff.lint]
# Rule groups to enable. Each letter/prefix is a category.
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes (unused imports, undefined names)
    "I",      # isort (import sorting)
    "B",      # flake8-bugbear (common bugs)
    "C4",     # flake8-comprehensions (better list/dict/set comprehensions)
    "UP",     # pyupgrade (modern Python syntax)
    "N",      # pep8-naming (naming conventions)
    "SIM",    # flake8-simplify (code simplification)
    "RUF",    # ruff-specific rules
]

# Rules to explicitly disable
ignore = [
    "E501",   # Line too long (formatter handles this)
]

[tool.ruff.lint.per-file-ignores]
# Tests can have some relaxed rules
"backend/tests/**/*.py" = [
    "N802",   # Function name should be lowercase (test functions often use descriptive names)
]

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"

# Indent with spaces (standard Python)
indent-style = "space"

# Use LF line endings (not CRLF) — Linux/Mac convention
line-ending = "lf"

[tool.mypy]
# Python version mypy assumes
python_version = "3.11"

# Where to look for code
files = ["backend/src"]

# ───── Strict mode flags ─────
# Each of these can be enabled individually; together they form "strict mode"
strict = true

# ───── Specific settings worth calling out ─────
# Warn about unused mypy "ignore" comments
warn_unused_ignores = true

# Warn when a function returns 'Any' (which silently disables type checking)
warn_return_any = true

# Warn about code after a return/raise that can never execute
warn_unreachable = true

# Show column numbers in error messages (helps pinpoint issues)
show_column_numbers = true

# Show error codes so we can selectively disable specific errors later if needed
show_error_codes = true

# Pretty, colorful output
pretty = true

# ───── Handling of external libraries ─────
# Some libraries (or older ones) don't have type stubs. Without this,
# mypy errors on every import of an untyped library.
# We DON'T set 'ignore_missing_imports = true' globally because that
# hides real problems. Instead, we silence specific libraries below as needed.

# Per-library overrides will be added here as we import third-party
# packages that lack type stubs. Example:
# [[tool.mypy.overrides]]
# module = ["some_untyped_library.*"]
# ignore_missing_imports = true

[tool.pytest.ini_options]
# Where pytest should look for tests
testpaths = ["backend/tests"]

# Match files starting with "test_"
python_files = ["test_*.py"]

# Match classes starting with "Test"
python_classes = ["Test*"]

# Match functions starting with "test_"
python_functions = ["test_*"]

# Enable async test support via pytest-asyncio
asyncio_mode = "auto"

# Show verbose output (individual test names)
addopts = "-v --strict-markers"
